requirements.txt
# Core FastAPI
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# Database
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
aiosqlite==0.20.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
cryptography==44.0.1

# Validation
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.8.0
email-validator==2.2.0

# Redis & Cache
redis==5.2.1
aioredis==2.0.1

# Rate Limiting
slowapi==0.1.9
limits==4.0.1

# Background Tasks
celery==5.4.0

# File Handling
aiofiles==24.1.0

# Cloud Services (Optional)
cloudinary==1.42.2

# HTTP Client
httpx==0.28.1

# Date & Time
python-dateutil==2.9.0.post0
pytz==2025.1

# Utils
Jinja2==3.1.5
PyYAML==6.0.2

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Database Async Driver
asyncpg==0.30.0
